Add skeleton for support of common isolated formatters
It is important that versions of formatting tools be locked across many
developers collaborating on a single project. This is often done using a
local installation of formatting tools via relevant package managers. An
example of this is prettier, which is rarely installed through system
package managers (it is not even available usually... except Arch of
course). The most common use case would be running the version of
prettier installed by npm, yarn, pnpm, etc. for the current project.